misc: move http package to root
- move pkg/httplog inside new http package
