+++ http/httplog/color.go
